+++ src/components/header-entry.tsx
@@ -33,7 +33,7 @@ function PageTitleDisplay({ pathname }: { pathname: string }) {
 
 function LogoDisplay() {
   return (
-    <Link href="/welcome" asChild>
+    <Link href="/" asChild>
       <TouchableOpacity>
         <LogoLightSmall />
       </TouchableOpacity>
